backend/pyproject.toml
[tool.poetry]
name = "liftdb"
version = "0.1.0"
description = ""
authors = ["Alex Kravetz <alexdkravetz@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^1.1.2"
SQLAlchemy = "^1.4.2"
Flask-SQLAlchemy = "^2.5.1"
Flask-Security-Too = "^4.0.0"
bcrypt = "^3.2.0"

[tool.poetry.dev-dependencies]
pylint = "^2.7.2"
pytest = "^6.2.2"
mypy = "^0.812"
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
